fix circular dependency between conf and utils
