+++ sparkmagic/sparkmagic/utils/configuration.py
@@ -8,8 +8,9 @@ from hdijupyterutils.configuration import override as _override
 from hdijupyterutils.configuration import override_all as _override_all
 from hdijupyterutils.configuration import with_override
 
-from .constants import HOME_PATH, CONFIG_FILE, MAGICS_LOGGER_NAME, LIVY_KIND_PARAM
-from .utils import get_livy_kind
+from .constants import HOME_PATH, CONFIG_FILE, MAGICS_LOGGER_NAME, LIVY_KIND_PARAM, \
+    LANG_SCALA, LANG_PYTHON, LANG_PYTHON3, LANG_R, \
+    SESSION_KIND_SPARKR, SESSION_KIND_SPARK, SESSION_KIND_PYSPARK, SESSION_KIND_PYSPARK3
 from sparkmagic.livyclientlib.exceptions import BadUserConfigurationException
 import sparkmagic.utils.constants as constants
 
@@ -28,6 +29,20 @@ def override_all(obj):
 
 _with_override = with_override(d, path)
 
+
+# Helpers
+
+def get_livy_kind(language):
+    if language == LANG_SCALA:
+        return SESSION_KIND_SPARK
+    elif language == LANG_PYTHON:
+        return SESSION_KIND_PYSPARK
+    elif language == LANG_PYTHON3:
+        return SESSION_KIND_PYSPARK3
+    elif language == LANG_R:
+        return SESSION_KIND_SPARKR
+    else:
+        raise BadUserConfigurationException("Cannot get session kind for {}.".format(language))
 
 # Configs
