+++ sparkmagic/sparkmagic/utils/utils.py
@@ -6,24 +6,9 @@ import pandas as pd
 import json
 from collections import OrderedDict
 
-from . import configuration as conf
+import sparkmagic.utils.configuration as conf
 import sparkmagic.utils.constants as constants
 from sparkmagic.livyclientlib.exceptions import BadUserDataException, DataFrameParseException
-from .constants import LANG_SCALA, LANG_PYTHON, LANG_PYTHON3, LANG_R, \
-    SESSION_KIND_SPARKR, SESSION_KIND_SPARK, SESSION_KIND_PYSPARK, SESSION_KIND_PYSPARK3
-
-
-def get_livy_kind(language):
-    if language == LANG_SCALA:
-        return SESSION_KIND_SPARK
-    elif language == LANG_PYTHON:
-        return SESSION_KIND_PYSPARK
-    elif language == LANG_PYTHON3:
-        return SESSION_KIND_PYSPARK3
-    elif language == LANG_R:
-        return SESSION_KIND_SPARKR
-    else:
-        raise ValueError("Cannot get session kind for {}.".format(language))
 
 
 def parse_argstring_or_throw(magic_func, argstring, parse_argstring=parse_argstring):
